created a data API for us to use
This setup.js file has all the data for each activity (and the
activities themselves). At the top I created handy functions for
returning data points we’ll commonly need to reference. If you need an
array of all the activity names, for example, all you have to do is
call App.allActivityNames. This saves you from the trouble of having to
parse the object/arrays for the information.