small fixes/footer design
I changed the footer so it’s “stuck” to the bottom of the browser page.
I’m totally open to displaying the footer differently, but it looked
odd just floating there in mid-air (on my large monitor).

The Images folder is buggy for me so I moved the header images into the
projectimages folder; I deleted the old images we aren’t using anymore.